Simplified tests on reflecting over the test context class and its fields.
It was necessary to remove tests driven with a fake SecurityManager, because changing what tests execute in this FieldExampleTest was causing these reflection-based tests to fail with BootstrapMethodError instead of the expected AccessControlException.

I'm not sure why exactly this was.  Perhaps the exclusion of a test caused the set of classes, constructors, and fields that were loaded before the axe-wielding NoReflectionForYouSecurityManager got a taste of its power?