Addressed unchecked conversion warnings by parameterizing NewJavaSpecRunner.gateway.
Now for an explanation of why parameterizing the gateway is a happy medium for abstraction and testability.

First off if the Runner and the Gateway were combined into one class, the only means of testing it would be to
make a bunch real classes that cause specific side effects, and then spying on those side effects.  While it
removes the difficulty of establishing a good, long-term Gateway interface, it's also quite painful to test.
So that option is out.

Having a gateway is a useful abstraction that separates concerns and makes both the runner and the gateway
easier to test.  However for the Gateway to do its job, it needs to get back the same concrete implementation
that it provides (ClassContext).  The Runner doesn't really care what kind of Context its using, but it must
declare its Gateway using a concrete parameter type in order to safely use the parameterized gateway.

That initially struck me as a smell of a leaky abstraction, but consider the fact that the Runner makes - as is
required by JUnit - its own dependencies.  There's no such thing as a JavaSpecRunner that works with a Gateway
parameterized by a type other than ClassContext.  So it's safe for it to declare this concrete type parameter.

Injecting a test double gateway with its own version of its Context classes is still possible, as long as the
test-specific Context classes extend ClassContext.  While I'd prefer to make ClassContext final until I have a
good reason not to, this small concession makes type-safe testing possible without having to make real
ClassContext objects.